Move module cbook.deprecation to _api.deprecation
