+++ lib/matplotlib/_api/deprecation.py
