+++ lib/matplotlib/cbook/__init__.py
@@ -29,7 +29,7 @@ import numpy as np
 
 import matplotlib
 from matplotlib import _c_internal_utils
-from .deprecation import (
+from matplotlib._api.deprecation import (
     deprecated, warn_deprecated,
     _rename_parameter, _delete_parameter, _make_keyword_only,
     _deprecate_method_override, _deprecate_privatize_attribute,
